style: arrange routes
