+++ routes/modules/api.js
@@ -1,8 +1,8 @@
 const express = require('express')
 const router = express.Router()
 
-const apiController = require('../../../controllers/pages/api-controller')
-const cpUpload = require('../../../middlewares/multer')
+const apiController = require('../../controllers/apis/api-controller')
+const cpUpload = require('../../middlewares/multer')
 
 router.get('/:id', apiController.getUserEditPage)
 router.post('/:id', cpUpload, apiController.postUserInfo)
