+++ routes/modules/user.js
@@ -1,14 +1,15 @@
 const express = require('express')
 const router = express.Router()
 
-const userController = require('../../../controllers/pages/user-controller')
-const cpUpload = require('../../../middlewares/multer')
+const userController = require('../../controllers/pages/user-controller')
+const cpUpload = require('../../middlewares/multer')
 
-router.get('/:id/tweets', userController.getUserTweets)
-router.get('/:id/followings', userController.getFollowings)
-router.get('/:id/followers', userController.getFollowers)
 router.get('/:id/likes', userController.getLikeTweets)
+router.get('/:id/tweets', userController.getUserTweets)
 router.get('/:id/replies', userController.getUserReplies)
+router.get('/:id/followers', userController.getFollowers)
+router.get('/:id/followings', userController.getFollowings)
+
 router.get('/:id', userController.getUserEditPage)
 router.post('/:id', cpUpload, userController.postUserInfo)
